Add the MICA output (cluster labels and UMAP/tSNE coordinates) to sparse eset object

This function is used to add the clustering results by MICA into the sparse eset object. Two types of resluts would be added to the phenoData slot of sparse eset object:

cluster ID of each cell

dimension reduction coordinates of each cell, either UMAP ot tSNE.

Usage

addMICAoutput(input_eset, mica_output_file, visual_method = "umap")

Arguments

input_eset

The sparse eset object to add the MICA output into

mica_output_file

The .txt file generated by MICA. It includes 4 columns: "ID" (Cell ID), "X" (UMAP_1 or tSNE_1), "Y" (UMAP_2 or tSNE_2), "label" (ClusterID)

visual_method

Character, method used for visualizing the clustering results: umap (the default) or tsne.

Value

A sparse eset object with clustering results added.

Examples

data(pbmc14k_expression.eset)
pbmc14k_log2cpm.eset <- addMICAoutput(pbmc14k_expression.eset,
                                      mica_output_file = system.file("extdata/demo_pbmc14k/MICA/clustering_UMAP_euclidean_20_2.05.txt", package = "scMINER"),
                                      visual_method = "umap")
#> Error in addMICAoutput(pbmc14k_expression.eset, mica_output_file = system.file("extdata/demo_pbmc14k/MICA/clustering_UMAP_euclidean_20_2.05.txt",     package = "scMINER"), visual_method = "umap"): The eset and the mica_output_file do not match. Please check and re-try.