Perform differential analysis between two groups

This function is used to perform the differential analysis between two groups. It supports there methods: "limma", "wilcoxon", and "t.test". This is an internal function for getDE() and getDA().

Usage

compare2groups(
  input_eset,
  group_by = "clusterID",
  g1 = NULL,
  g0 = NULL,
  use_method = "limma"
)

Arguments

input_eset

The expression set object that filtered, normalized and log-transformed

group_by

Character, name of the column for grouping, usually the column of cell types or clusters. Default: "clusterID".

g1

A vector of character defining the fore-ground group or NULL. Default: NULL.

g0

A vector of character defining the back-ground group or NULL. Default: NULL.

use_method

Character, method used for differential analysis: "limma" (the default), "wilcoxon", and "t.test".

Value

A data frame. Rows are genes/drivers, and columns are 11 statistics of differential analysis.

Examples

## to call this function
data(pbmc14k_expression.eset)
res <- compare2groups(input_eset = pbmc14k_expression.eset,
                      group_by = "cellType",
                      g1 = c("CD4TCM", "CD4TN", "CD4Treg", "CD8TN", "Monocyte", "NK"),
                      g0 = c("B"),
                      use_method = "limma")
#> 	 2999 cells were found for g1.
#> 	 501 cells were found for g0.