Create a project space for scMINER analysis

This function is used to create a folder of the specified project names in the specified project directory as the project space to run scMINER analysis. It also creastes 4 subfolers inside of it: "DATA", "MICA", "SJARACNe" and "PLOT".

Usage

createProjectSpace(project_dir, project_name, do.unlink = FALSE)

Arguments

project_dir

The directory to create the project space in

project_name

The name of the project, will be used as the name of the folder

do.unlink

Logical, whether to remove the files and/or folders inside of pre-existing project space. Default: FALSE.

Value

It creates a folder of project name and 4 subfolders in the project directory, and returns the path of project space.

Examples

if (FALSE) { # \dontrun{
scminer_dir <- createProjectSpace(project_dir = "path-to-a-folder",
                                  project_name = "PBMC14k")
} # }