Generate a quality control report from sparse eset object

This function is used to generate a html quality control report from a sparse eset object. Compared with the summary table return by filterSparseEset(), the output report contains more comprehensive and detailed QC results and can be used to estimate the cutoffs to filter the eset object. It also contains some plots for presentation purpose.

Usage

drawSparseEsetQC(
  input_eset,
  output_html_file,
  overwrite = FALSE,
  group_by = NULL
)

Arguments

input_eset

The sparse eset object for quality control analysis

output_html_file

The path of the output .html file

overwrite

Logical, whether to overwrite the output .html file if it already exists. Default: FALSE.

group_by

Character or NULL, Name of column in pData(eset) used for grouping. Default: NULL.

Value

A html-formatted quality control report of sparse eset object

Examples

## 1. To generate the QC report in a group-specific manner, recommended whenever group information is avaiable.
drawSparseEsetQC(input_eset = pbmc14k_raw.eset,
                 output_html_file = "/your-path/PBMC14k/PLOT/pbmc14k_rawCount.html",
                 overwrite = FALSE,
                 group_by = "trueLabel")
#> Checkinig the output html file ...
#> Warning: cannot create dir '/your-path/PBMC14k', reason 'No such file or directory'
#> 	The directory of output html file does not exists. Now it has been created!
#> Checkinig the group_by information in eset ...
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'pbmc14k_raw.eset' not found

## 2. To generate the QC report from a whole view
drawSparseEsetQC(input_eset = pbmc14k_raw.eset,
                 output_html_file = "/your-path/PBMC14k/PLOT/pbmc14k_rawCount.html",
                 overwrite = FALSE,
                 group_by = NULL)
#> Checkinig the output html file ...
#> Warning: cannot create dir '/your-path/PBMC14k', reason 'No such file or directory'
#> 	The directory of output html file does not exists. Now it has been created!
#> Checking the availability of the 5 metrics ('nCell', 'nUMI', 'nFeature', 'pctMito', 'pctSpikeIn') used for quanlity control ...
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'fData': object 'pbmc14k_raw.eset' not found