Filter the cells and/or features of sparse eset object using automatic or self-customized cutoffs

This function is used to remove the cells and features of low quality. It provides two modes to define the cutoffs:

"auto": in this mode, scMINER will estimate the cutoffs based on Median ± 3*MAD (maximum absolute deviation). This mode works well for the matrix of raw UMI counts or TPM (Transcripts Per Million) values.

"manual": in this mode, the users can manually specify the cutoffs, both low and high, of all 5 metrics: nUMI, nFeature, pctMito, pctSpikeIn for cells, and nCell for genes. No cells or features would be removed under the default cutoffs of each metrics.

Usage

filterSparseEset(
  input_eset,
  filter_mode = "auto",
  filter_type = "both",
  gene.nCell_min = NULL,
  gene.nCell_max = NULL,
  cell.nUMI_min = NULL,
  cell.nUMI_max = NULL,
  cell.nFeature_min = NULL,
  cell.nFeature_max = NULL,
  cell.pctMito_min = NULL,
  cell.pctMito_max = NULL,
  cell.pctSpikeIn_min = NULL,
  cell.pctSpikeIn_max = NULL
)

Arguments

input_eset

The sparse eset object to be filtered

filter_mode

Character, mode to apply the filtration cutoffs: "auto" (the default) or "manual".

filter_type

Character, objective type to be filtered: "both" (the default), "cell" or "feature" .

gene.nCell_min

Numeric, the minimum number of cells that the qualified genes are identified in. Default: 1.

gene.nCell_max

Numeric, the maximum number of cells that the qualified genes are identified in. Default: Inf.

cell.nUMI_min

Numeric, the minimum number of total UMI counts per cell that the qualified cells carry. Default: 1.

cell.nUMI_max

Numeric, the maximum number of total UMI counts per cell that the qualified cells carry. Default: Inf.

cell.nFeature_min

Numeric, the minimum number of non-zero Features per cell that the qualified cells carry. Default: 1.

cell.nFeature_max

Numeric, the maximum number of non-zero Features per cell that the qualified cells carry. Default: Inf.

cell.pctMito_min

Numeric, the minimum percentage of UMI counts of mitochondrial genes that the qualified cells carry. Default: 0.

cell.pctMito_max

Numeric, the maximum percentage of UMI counts of mitochondrial genes that the qualified cells carry. Default: 1.

cell.pctSpikeIn_min

Numeric, the minimum percentage of UMI counts of spike-in that the qualified cells carry. Default: 0.

cell.pctSpikeIn_max

Numeric, the maximum percentage of UMI counts of spike-in that the qualified cells carry. Default: 1.

Value

A filtered sparse eset object. It also prints the summary of filtration to the screen.

Examples

filtered.eset <- filterSparseEset(raw.eset) ## filter the input eset using the cutoffs calculated by scMINER.
#> Checking the availability of the 5 metrics ('nCell', 'nUMI', 'nFeature', 'pctMito', 'pctSpikeIn') used for filtration ...
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'fData': object 'raw.eset' not found
filtered.eset <- filterSparseEset(raw.eset, gene.nCell_min = 10, cell.nUMI_min = 500, cell.nFeature_min = 100, cell.nFeature_max = 5000, cell.pctMito_max = 0.15)
#> Checking the availability of the 5 metrics ('nCell', 'nUMI', 'nFeature', 'pctMito', 'pctSpikeIn') used for filtration ...
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'fData': object 'raw.eset' not found