Prepare the standard input files for scMINER Portal

This function is used to generated the standard input files that can be directly uploaded to scMINER Portal for visualization and beyond.

Usage

generatePortalInputs(
  input_expression.eset = NULL,
  input_expression.seuratObj = NULL,
  group_by = NULL,
  input_activity.eset = NULL,
  input_network.dir = NULL,
  input_network.table = NULL,
  output_dir = NULL
)

Arguments

input_expression.eset

A eset object of expression data which has been filerted, normalized and log-transformed. Default: NULL.

input_expression.seuratObj

A Seurat object with reduction results (umap and/or tsne). Default: NULL.

group_by

Character, name of the column in phenoData of eset object or meta.data of Seurat Obj that defines the groups for colorcoding in scMINER Portal. Default: NULL.

input_activity.eset

A eset object of activity data calculated by scMINER. Default: NULL.

input_network.dir

Character, the path to the SJARACNe directory. Default: NULL.

input_network.table

A table of network information, it contains at least three column: "CellGroup" (name of groups), "NetworkType" (type of network, TF or SIG) and "NetworkFile" (path to network files). Default: NULL.

output_dir

Character, the path to the output directory. Default: NULL.

Value

This function generated 1-3 standard files which can be uploaded to scMINER Portal directly.

Examples

## 1. the most commonly used command
generatePortalInputs(input_expression.eset = expression_clustered.eset, group_by = "cellType", input_activity.eset = activity_clustered.eset, input_network.dir = "./SJARACNe", output_dir = "./scMINERportal")
#> The output directory has been created:  ./scMINERportal .
#> Error in eval(expr, envir, enclos): object 'expression_clustered.eset' not found

## 2. prepare expression data from Seurat object ("pbmc14.obj")
generatePortalInputs(input_expression.seuratObj = pbmc14.obj, output_dir = "./path-to-output_dir")
#> The output directory has been created:  ./path-to-output_dir .
#> Error in eval(expr, envir, enclos): object 'pbmc14.obj' not found

## 3. prepare network data from a table
network.table <- data.frame(CellGroup = c("CD4__CD25_T_Reg", "CD4__CD25_T_Reg", "CD19__B", "CD19__B"),NetworkType = c("SIG", "TF", "SIG", "TF"),
                            NetworkFile = c("./sjaracne/CD4__CD25_T_Reg/SIG/b100_pce-3/sjaracne_workflow-1474c41b-067b-4f86-ab99-09f73dadb16g/consensus_network_ncol_.txt",
                                            "./sjaracne/CD4__CD25_T_Reg/TF/b100_pce-3/sjaracne_workflow-a93cd6db-7253-4ffb-ae4e-633b9dedf11d/consensus_network_ncol_.txt",
                                            "./sjaracne/CD19__B/SIG/sjaracne_workflow-da0c3c72-7afb-44fa-973b-e4d767e20b6f/consensus_network_ncol_.txt",
                                            "./sjaracne/CD19__B/TF/sjaracne_workflow-0426ea12-10bf-428c-b199-d5bd1a7aab5f/consensus_network_ncol_.txt"))
generatePortalInputs(input_expression.eset = expression_clustered.eset, group_by = "cellType", input_network.table = network.table, output_dir = "./path-to-output_dir")
#> Error in eval(expr, envir, enclos): object 'expression_clustered.eset' not found