Calculate driver activities in batch from the SJARACNe directory

This function is used to calculate the driver activities of multiple groups from a scMINER directory. To calculate driver activities of one single group, please use getActivity_individual().

Usage

getActivity_inBatch(
  input_eset,
  sjaracne_dir,
  group_name,
  group_exclude = NULL,
  network_tag.tf = NULL,
  network_tag.sig = NULL,
  driver_type = "TF_SIG",
  activity_method = "mean",
  do.z_normalization = TRUE
)

Arguments

input_eset

The expression set object which has been filtered, normalized and log-transformed

sjaracne_dir

The path to the SJARACNe directory

group_name

Character, name of the column for grouping, usually the column of cell types or clusters

group_exclude

A vector of group names to exclude in activity calculation

network_tag.tf

Character or NULL, the tag used to distinguish different SJARACNe runs from the same input files. This is usually the name of the folder(s) in "TF" folder of each group. Default: NULL.

network_tag.sig

Character or NULL, the tag used to distinguish different SJARACNe runs from the same input files. This is usually the name of the folder(s) in "SIG" folder of each group. Default: NULL.

driver_type

Character, type of the pre-defined driver list to use. Should be one of: "TF" for transcriptional factors only, "SIG" for signaling genes only, or "TF_SIG" for both. Default: "TF_SIG".

activity_method

Character, method used to calculate the activity: "mean" (the default), "weightedmean", "absmean" or "maxmean".

do.z_normalization

Logical, whether to do the z-normalization on the gene expression values in each sample. Set if to FALSE only when the expression values has been scaled in each cell. Default: TRUE.

Value

A expression set object. The assayData is the activity matrix of all cells of all groups, drivers by cells. The phenoData and featureData are exactly save with the input eset.

Examples

## 1. when no tag was used in runing SJARACNE: the network file folder ("sjaracne_workflow-*") is directly under TF/SIG folder of each group.
activity.eset <- getActivity_inBatch(input_eset = normalized.eset, sjaracne_dir = "./SJARACNe", group_name = "cell_type", driver_type = "TF_SIG", activity_method = "mean", do.z_normalization = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'normalized.eset' not found

## 2. when tag (e.g. "bs_100" ) was used: the nework file folder ("sjaracne_workflow-*") is directly under a subfolder "bs_100" of the TF/SIG folder of each group.
activity.eset <- getActivity_inBatch(input_eset = normalized.eset, sjaracne_dir = "./SJARACNe", group_name = "cell_type", network_tag.tf = "bs_100", network_tag.sig = "bs_100", driver_type = "TF_SIG", activity_method = "mean", do.z_normalization = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'normalized.eset' not found

## 3. to calculate the activities of TF only
activity.eset <- getActivity_inBatch(input_eset = normalized.eset, sjaracne_dir = "./SJARACNe", group_name = "cell_type", network_tag.tf = "bs_100", network_tag.sig = "bs_100", driver_type = "TF", activity_method = "mean", do.z_normalization = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'normalized.eset' not found

## 4. to exclude some groups in the activity calculation (e.g. "NK" and "Monocyte")
activity.eset <- getActivity_inBatch(input_eset = normalized.eset, sjaracne_dir = "./SJARACNe", group_name = "cell_type", group_exclude = c("NK", "Monocyte"), network_tag.tf = "bs_100", network_tag.sig = "bs_100", driver_type = "TF", activity_method = "mean", do.z_normalization = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'normalized.eset' not found

## 5. when calculate the activities from the gene expression values scaled by other methods (e.g. ScaleData() from Seurat package)
activity.eset <- getActivity_inBatch(input_eset = normalized.eset, sjaracne_dir = "./SJARACNe", group_name = "cell_type", network_tag.tf = "bs_100", network_tag.sig = "bs_100", driver_type = "TF_SIG", activity_method = "mean", do.z_normalization = FALSE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'normalized.eset' not found