Calculate driver activities per group from network files

This function is used to calculate the driver activities of one single group from the sparse eset obj and networks generated by SJARACNe. To calculate driver activities of multiple groups from a scMINER directory, please use getActivity_inBatch().

Usage

getActivity_individual(
  input_eset,
  network_file.tf = NULL,
  network_file.sig = NULL,
  driver_type = "TF_SIG",
  activity_method = "mean",
  do.z_normalization = TRUE
)

Arguments

input_eset

The group-specific expression set object which has been filtered, normalized and log-transformed

network_file.tf

The path to the TF network file generated by SJARACNe

network_file.sig

The path to the SIG network file generated by SJARACNe

driver_type

Character, type of the pre-defined driver list to use: "TF" for transcriptional factors only, "SIG" for signaling genes only, or "TF_SIG" for both. Default: "TF_SIG".

activity_method

Character, method used to calculate the activity: "mean" (the default), "weightedmean", "absmean" or "maxmean".

do.z_normalization

Logical, whether to do the z-normalization on the gene expression values in each sample. Set if to FALSE only when the expression values has been scaled in each cell. Default: TRUE.

Value

A expression set object of the group-of-interest. The assayData is the activity matrix of all cells of all groups, drivers by cells. The phenoData and featureData are exactly save with the input eset.

Examples

activity_group.eset <- getActivity_individual(input_eset = group_specific.est, network_file.tf = "consensus_network_ncol_.txt", network_file.sig = "consensus_network_ncol_.txt", driver_type = "TF_SIG")
#> Warning: cannot open file 'consensus_network_ncol_.txt': No such file or directory
#> Error in file(file, "rt"): cannot open the connection