Perform differential expression analysis on expression set object

This function is used to perform the differential expression analysis on sparse eset object. It supports there methods: "limma", "wilcoxon", and "t.test".

Usage

getDE(
  input_eset,
  group_by = "clusterID",
  g1 = NULL,
  g0 = NULL,
  use_method = "limmma"
)

Arguments

input_eset

The expression set object that filtered, normalized and log-transformed

group_by

Character, name of the column for grouping, usually the column of cell types or clusters. Default: "clusterID".

g1

A vector of character defining the fore-ground group or NULL. Default: NULL.

g0

A vector of character defining the back-ground group or NULL. Default: NULL.

use_method

Character, method used for differential analysis: "limma" (the default), "wilcoxon", and "t.test".

Value

A data frame. Rows are genes/drivers, and columns are 11 statistics of differential analysis.

Examples

## 1. To perform differential expression analysis in a 1-vs-rest manner for all groups in "clusterID" column
de_res <- getDE(input_eset = clustered.eset,
                group_by = "clusterID",
                use_method = "limma")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'clustered.eset' not found

## 2. To perform differential expression analysis in a 1-vs-rest manner for one specific group in "clusterID" column
de_res <- getDE(input_eset = clustered.eset,
                group_by = "clusterID",
                g1 = c("1"),
                use_method = "limma")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'clustered.eset' not found

## 3. To perform differential expression analysis in a rest-vs-1 manner for one specific group in "clusterID" column
de_res <- getDE(input_eset = clustered.eset,
                group_by = "clusterID",
                g0 = c("1"),
                use_method = "limma")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'clustered.eset' not found

## 4. To perform differential expression analysis in a 1-vs-1 manner for groups in "clusterID" column
de_res <- getDE(input_eset = clustered.eset,
                group_by = "clusterID",
                g1 = c("1"),
                g0 = c("3"),
                use_method = "limma")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'pData': object 'clustered.eset' not found