SparseEset object of PBMC14k dataset

This dataset contains the SparseEset object of PBMC14k dataset. For demonstration purposes, it has been downsampled to 3.5k cells, with 500 cells per population.

Usage

pbmc14k_expression.eset

Format

pbmc14k_expression.eset

A large dgCMatrix with 17,986 rows and 14,000 columns:

This data set provides the SparseEset object of PBMC14k dataset that has been filtered, normalized, clustered and annotated.

Source

It's generated by scMINER from Filtered_DownSampled_SortedPBMC_data.csv from https://zenodo.org/record/3357167#.YhQNF2RKj6V