Raw count matrix of PBMC14k dataset

This dataset contains the raw UMI counts of 14,000 Peripheral Blood Mononuclear Cells (PBMCs) of 7 known cell types, with 2,000 cell each.

Usage

pbmc14k_rawCount

Format

pbmc14k_rawCount

A large dgCMatrix with 17,986 rows and 14,000 columns:

CACTTTGACGCAAT

UMI counts of all genes of the cell "CACTTTGACGCAAT"

GTTACGGAAACGAA

UMI counts of all genes of the cell "GTTACGGAAACGAA"

AGTCACGACAGGAG

UMI counts of all genes of the cell "AGTCACGACAGGAG"

...

Source

A subset of Filtered_DownSampled_SortedPBMC_data.csv from https://zenodo.org/record/3357167#.YhQNF2RKj6V