Assess the quality of each network generated by SJARACNe

This function is used to assess the quality of networks generated by SJARACNe. It returns a summary table of key statistics of networks. The users can also generate the html quality control report by turning generate_html = TRUE.

Usage

drawNetworkQC(
  network_file = NULL,
  sjaracne_dir = NULL,
  directed = TRUE,
  weighted = TRUE,
  generate_html = TRUE,
  outdir = NULL,
  prefix = NULL
)

Arguments

network_file

The path to the network file (consensus_network_ncol_.txt) for quality control

sjaracne_dir

The path to the folder of SJARACNe runs. When this argument is given, scMINER will automatically retrieve all network files (consensus_network_ncol_.txt) in this folder and generate quality control report for all of them.

directed

Logical, whether the network is directed or not. All networks generated by SJARACNe are directed. Default: TRUE.

weighted

Logical, whether the edges of the network is weighted or not. All networks generated by SJARACNe are weighted by mutual information ("MI" column). Default: TRUE.

generate_html

Logical, whether to generate a html report. Default: TRUE.

outdir

Character or NULL, the path to save the html report. If NULL, the html report will be saved in the same folder of the network file. Default: NULL. Ignored if generate_html = FALSE.

prefix

Character or NULL, the character string to add in front of the html report file name. Default: NULL. Ignored if generate_html = FALSE.

Value

This function will print the statistics of several key quality metrics of network(s). If generate_html is set TRUE, it also generates a html file of quality control report and save it to the same folder of the network file (by default) or the folder specified by "outdir".

Examples

## 1. assess the quality of network from a network file
drawNetworkQC(network_file = ./SJARACNE/B/SIG/b100/consensus_network_ncol_.txt, generate_htmal = TRUE) # the html file will be saved to the same folder as the network file
#> Error in drawNetworkQC(network_file = ./SJARACNE/B/SIG/b100/consensus_network_ncol_.txt,     generate_htmal = TRUE): unused argument (generate_htmal = TRUE)
drawNetworkQC(network_file = ./SJARACNE/B/SIG/b100/consensus_network_ncol_.txt, generate_htmal = TRUE, outdir = "/path-to-cutomized-folder") # the html file will be saved to the customized folder
#> Error in drawNetworkQC(network_file = ./SJARACNE/B/SIG/b100/consensus_network_ncol_.txt,     generate_htmal = TRUE, outdir = "/path-to-cutomized-folder"): unused argument (generate_htmal = TRUE)
drawNetworkQC(network_file = ./SJARACNE/B/SIG/b100/consensus_network_ncol_.txt, generate_htmal = TRUE, prefix = "PBMC14") # add the "PBMC14" in front of the default name of html report
#> Error in drawNetworkQC(network_file = ./SJARACNE/B/SIG/b100/consensus_network_ncol_.txt,     generate_htmal = TRUE, prefix = "PBMC14"): unused argument (generate_htmal = TRUE)

## 2. assess the qulity of network from a directory
drawNetworkQC(sjaracne_dir = ./SJARACNE, generate_html = TRUE) # the html file will be saved to the same folder as the network file
#> Error in eval(expr, envir, enclos): object '.' not found