Normalize and log-transform the sparse eset object

This function is used to normalize and log-transform the sparse eset object. The default method is "log21p".

Usage

normalizeSparseEset(
  input_eset,
  scale_factor = 1e+06,
  do.logTransform = TRUE,
  log_base = 2,
  log_pseudoCount = 1
)

Arguments

input_eset

The sparse eset object for normalization

scale_factor

Numeric, the library size to normalize to. Default: 1000000.

do.logTransform

Logical, whether to do log-transformation. Default: TRUE.

log_base

Numeric, the base of log-transformation. Usually 2, exp(1) or 10. Default: 2.

log_pseudoCount

Numeric, the pseudo count to add to avoid "-Inf" in log-transformation. Default: 1.

Value

A sparse eset object that has been normalized and log-transformed

Examples

pbmc14k_log2cpm.eset <- normalizeSparseEset(pbmc14k_filtered.eset,
                                            scale_factor = 1000000,
                                            log_base = 2,
                                            log_pseudoCount = 1)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'exprs': object 'pbmc14k_filtered.eset' not found